refactor(gui): Phase 5 - split AppSignalHandler into by-concern partials
The ImageRendererBase treatment: the ctor's subscription groups move verbatim
into per-concern Subscribe* methods, one partial per area -- .Planner (search
input + schedule building), .SkyMap (F3 search + map signals + SkyMapViewingUtc),
.Equipment (text inputs + device action signals + RunDeviceOpOffRenderThreadAsync),
.LiveSession (session + preview-mode signals), .Polar, .Flats. Core file keeps
the shared scaffolding (fields, ctor wiring, planner init, telemetry polls,
routing helpers): 2566 -> 858 lines.

Mechanics chosen for zero behaviour change:
- Each Subscribe* method opens with `var x = _x;` field aliases so the moved
  closures' captured-parameter names keep resolving -- handler bodies are
  byte-identical (ctor-body and method-body statements share the indent depth).
- The ctor calls the groups in the original order; SignalBus invokes subscribers
  in registration order, so the call order is load-bearing (commented as such).
- The two cross-section couplings are promoted: the `autoCompleteCache` ctor
  local becomes the `_autoCompleteCache` field (Planner partial); `shutdownToken`
  becomes a SubscribePreview parameter. CommitSuggestion stays a local function
  inside SubscribePlannerSearch (its only caller).

Organizational only, no lines removed. Verified: 0-warning build (incl. GUI
Release + packages, CI parity), 80 targeted signal-handler tests, full unit
suite 3344 passed / 0 failed.
