+++ docs/todo/infra.md
@@ -29,7 +29,7 @@ Part of the TianWen TODO set. See [TODO.md](../../TODO.md) for the index and the
     - [x] `SkyMapSlewToObjectSignal` — largely already routing (two `MountActions` calls); the two-click Sun-slew confirmation state machine → testable `GuiAppState.GateSunSlew`.
     - [x] TextInput commit callbacks — `saveSite` parse/validate → `EquipmentActions.TryParseSite`, mount push → `EquipmentActions.PushSiteToMountIfProfileWinsAsync`; `StringSettingInput.OnCommit` masked-secret/URI decision → `EquipmentActions.CommitDeviceSetting`. (`ProfileName`/`GuiderFL`/`saveOta` left as-is: already thin — single helper call or single-field set + save.)
   - Pinned by `RouteOnlyExtractionTests` (`TryParseSite`, `CommitDeviceSetting`, `GateSunSlew`).
-- [x] **Signal-handler boilerplate reduction** -- DONE (Phases 1-3, branch `refactor/signal-handler-boilerplate`): `Notify` / guard helpers (`EnsureSessionIdle` / `TryGetConnected<T>` / `TryResolveIdleOtaFocuser`) / `RunTracked` (over the upstreamed `DIR.Lib.BackgroundTaskTracker.RunGuarded`) swept across the handlers and extended to the ones added since the draft (Flats/manual-cover/comets). Bespoke error sites left untouched; Tier 4 `Wire<T>` + Phase 5 partial split deferred. `AppSignalHandler.cs` ~2687 -> ~2566 lines. [docs/plans/signal-handler-boilerplate.md](../plans/signal-handler-boilerplate.md).
+- [x] **Signal-handler boilerplate reduction** -- DONE (Phases 1-3 + 5, branch `refactor/signal-handler-boilerplate`): `Notify` / guard helpers (`EnsureSessionIdle` / `TryGetConnected<T>` / `TryResolveIdleOtaFocuser`) / `RunTracked` (over the upstreamed `DIR.Lib.BackgroundTaskTracker.RunGuarded`) swept across the handlers and extended to the ones added since the draft (Flats/manual-cover/comets); the ctor's subscription groups then split verbatim into per-concern `Subscribe*` partials (`.Planner/.SkyMap/.Equipment/.LiveSession/.Polar/.Flats.cs`, call order = registration order). Bespoke error sites left untouched; Tier 4 `Wire<T>` dropped by its own kill criterion. Core file ~2687 -> ~860 lines. [docs/plans/signal-handler-boilerplate.md](../plans/signal-handler-boilerplate.md).
 - [ ] **OnStep follow-ups** (leftover from the OnStep commit series):
   - [ ] MoveAxis via `:Mn/:Ms/:Me/:Mw#` + `:Qe/Qw/Qn/Qs#` + `:RA/:RE` rates — enables direct jog buttons in GUI
   - [ ] Per-axis guide-rate setter via `:Rn#` (index 0–9) + `:GX90#` query — enables `CanSetGuideRates = true` on the OnStep override
