+++ src/TianWen.UI.Abstractions/AppSignalHandler.Planner.cs
@@ -1,0 +1,185 @@
+using System;
+using System.Collections.Concurrent;
+using System.Collections.Generic;
+using System.Collections.Immutable;
+using System.Diagnostics.CodeAnalysis;
+using System.Globalization;
+using System.IO;
+using System.Linq;
+using System.Threading;
+using System.Threading.Tasks;
+using DIR.Lib;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Logging;
+using TianWen.Lib.Astrometry;
+using TianWen.Lib.Astrometry.Catalogs;
+using TianWen.Lib.Astrometry.Comets;
+using TianWen.Lib.Astrometry.PlateSolve;
+using TianWen.Lib.Astrometry.SOFA;
+using TianWen.Lib.Devices;
+using TianWen.Lib.Devices.Fake;
+using TianWen.Lib.Devices.Guider;
+using TianWen.Lib.Devices.Weather;
+using TianWen.Lib.Extensions;
+using TianWen.Lib.Imaging;
+using TianWen.Lib.Sequencing;
+using TianWen.Lib.Sequencing.PolarAlignment;
+
+namespace TianWen.UI.Abstractions
+{
+    // AppSignalHandler.Planner.cs -- planner search input callbacks + schedule building.
+    // One partial per concern (see the class doc in AppSignalHandler.cs); handler bodies
+    // moved verbatim from the single-file ctor in the Phase-5 by-area split.
+    public partial class AppSignalHandler
+    {
+        /// <summary>Autocomplete candidates for the planner search box; set by the host
+        /// after catalog load via <see cref="SetAutoCompleteCache"/>. Was a ctor local
+        /// shared with the search closures before the by-area split.</summary>
+        private string[]? _autoCompleteCache;
+
+        /// <summary>Wires the planner search input callbacks (autocomplete, suggestion commit).</summary>
+        private void SubscribePlannerSearch(SignalBus bus)
+        {
+            // Aliases over the injected fields keep the moved handler bodies verbatim
+            // (the closures captured the ctor's parameters before the by-area split).
+            var appState = _appState;
+            var plannerState = _plannerState;
+            var sp = _sp;
+
+            // ---------------------------------------------------------------
+            // Wire planner search input callbacks
+            // ---------------------------------------------------------------
+
+            plannerState.SearchInput.OnCommit = text =>
+            {
+                plannerState.Suggestions.Clear();
+                plannerState.SuggestionIndex = -1;
+                plannerState.LastSuggestionQuery = "";
+
+                if (appState.ActiveProfile is not null && text.Length > 0)
+                {
+                    var transform = TransformFactory.FromProfile(appState.ActiveProfile, _timeProvider, out _);
+                    if (transform is not null)
+                    {
+                        var db = sp.GetRequiredService<TianWen.Lib.Astrometry.Catalogs.ICelestialObjectDB>();
+                        var resultIdx = PlannerActions.SearchTargets(plannerState, db, transform, text, plannerState.Comets);
+                        if (resultIdx >= 0)
+                        {
+                            plannerState.SelectedTargetIndex = resultIdx;
+                            OnPlannerEnsureVisible?.Invoke(resultIdx);
+                        }
+                    }
+                }
+                return Task.CompletedTask;
+            };
+
+            plannerState.SearchInput.OnCancel = () =>
+            {
+                plannerState.SearchInput.Clear();
+                plannerState.SearchResults = [];
+                plannerState.Suggestions.Clear();
+                plannerState.SuggestionIndex = -1;
+                plannerState.LastSuggestionQuery = "";
+                bus.Post(new DeactivateTextInputSignal());
+                plannerState.NeedsRedraw = true;
+            };
+
+            plannerState.SearchInput.OnTextChanged = text =>
+            {
+                if (_autoCompleteCache is not null)
+                {
+                    PlannerActions.UpdateSuggestions(plannerState, _autoCompleteCache, text);
+                }
+            };
+
+            // Autocomplete navigation: Up/Down/Return/Escape when suggestions are visible
+            plannerState.SearchInput.OnKeyOverride = key =>
+            {
+                if (plannerState.Suggestions.Count == 0)
+                {
+                    return false;
+                }
+
+                switch (key)
+                {
+                    case TextInputKey.Backspace or TextInputKey.Delete:
+                        return false; // Let the text input handle it, OnTextChanged will update suggestions
+
+                    case TextInputKey.Enter when plannerState.SuggestionIndex >= 0:
+                        CommitSuggestion(plannerState.Suggestions[plannerState.SuggestionIndex]);
+                        return true;
+
+                    case TextInputKey.Escape:
+                        plannerState.Suggestions.Clear();
+                        plannerState.SuggestionIndex = -1;
+                        plannerState.LastSuggestionQuery = "";
+                        appState.NeedsRedraw = true;
+                        return true;
+
+                    default:
+                        return false;
+                }
+            };
+
+            // Local helper captured by the search-input closures above.
+            void CommitSuggestion(string suggestion)
+            {
+                plannerState.SearchInput.Text = suggestion;
+                plannerState.SearchInput.CursorPos = suggestion.Length;
+                plannerState.Suggestions.Clear();
+                plannerState.SuggestionIndex = -1;
+                plannerState.LastSuggestionQuery = suggestion;
+
+                if (appState.ActiveProfile is not null)
+                {
+                    var transform = TransformFactory.FromProfile(appState.ActiveProfile, _timeProvider, out _);
+                    if (transform is not null)
+                    {
+                        var db = sp.GetRequiredService<TianWen.Lib.Astrometry.Catalogs.ICelestialObjectDB>();
+                        var resultIdx = PlannerActions.CommitSuggestion(plannerState, db, transform, suggestion, plannerState.Comets);
+                        if (resultIdx >= 0)
+                        {
+                            plannerState.SelectedTargetIndex = resultIdx;
+                            OnPlannerEnsureVisible?.Invoke(resultIdx);
+                        }
+                    }
+                }
+                appState.NeedsRedraw = true;
+            }
+        }
+
+        /// <summary>Wires schedule building (shared between planner preview and session start).</summary>
+        private void SubscribeScheduleBuilding(SignalBus bus)
+        {
+            // Aliases over the injected fields keep the moved handler bodies verbatim
+            // (the closures captured the ctor's parameters before the by-area split).
+            var appState = _appState;
+            var plannerState = _plannerState;
+            var sessionState = _sessionState;
+
+            // ---------------------------------------------------------------
+            // Schedule building (shared between planner preview and session start)
+            // ---------------------------------------------------------------
+            bus.Subscribe<BuildScheduleSignal>(signal =>
+            {
+                if (appState.ActiveProfile is not { } profile) return;
+                var transform = TransformFactory.FromProfile(profile, _timeProvider, out _);
+                if (transform is null) return;
+
+                var profileData = profile.Data ?? ProfileData.Empty;
+                var (filters, design) = GetFirstOtaFilterConfig(profileData);
+                // Use the SAME effective defaults the session-start path uses so the planner
+                // PREVIEW matches what the session will actually capture (it previously hardcoded
+                // gain=120/offset=10/sub=120s here, diverging from both the config and the
+                // f-ratio exposure shown on the target rows). Gain/offset stay null so the per-OTA
+                // camera settings drive them, exactly as StartSession does.
+                PlannerActions.BuildSchedule(plannerState, sessionState, transform,
+                    defaultGain: null, defaultOffset: null,
+                    defaultSubExposure: SessionContent.EffectiveDefaultSubExposure(sessionState),
+                    defaultObservationTime: TimeSpan.FromMinutes(60),
+                    availableFilters: filters,
+                    opticalDesign: design);
+            });
+        }
+    }
+}
