+++ src/TianWen.UI.Abstractions/AppSignalHandler.cs
@@ -40,8 +40,15 @@ namespace TianWen.UI.Abstractions
     /// Sync subscribers run inline on the render thread; async subscribers are submitted
     /// to <see cref="BackgroundTaskTracker"/> (never fire-and-forget).
     /// </para>
+    /// <para>
+    /// Split by concern across partial files (the ImageRendererBase treatment): this file
+    /// holds the shared scaffolding (fields, ctor wiring, planner init, telemetry polls,
+    /// routing helpers); the subscription groups live in <c>AppSignalHandler.Planner.cs</c> /
+    /// <c>.SkyMap.cs</c> / <c>.Equipment.cs</c> / <c>.LiveSession.cs</c> / <c>.Polar.cs</c> /
+    /// <c>.Flats.cs</c>.
+    /// </para>
     /// </summary>
-    public class AppSignalHandler
+    public partial class AppSignalHandler
     {
         private readonly IServiceProvider _sp;
         private readonly GuiAppState _appState;
@@ -80,21 +87,6 @@ namespace TianWen.UI.Abstractions
                 : OpticalDesign.Unknown;
             return (filters is { Count: > 0 } ? filters : null, design);
         }
-
-        /// <summary>
-        /// Runs a device connect/disconnect on a thread-pool thread so its <b>synchronous prefix
-        /// never executes on the render thread</b>. <see cref="SignalBus.ProcessPending"/> invokes
-        /// async signal handlers inline on the render thread, running each handler up to its first
-        /// yielding <c>await</c> before the returned task is handed to the tracker. Several drivers
-        /// block synchronously <i>before</i> that first await — most notably ASCOM COM drivers whose
-        /// <c>Connected = true/false</c> setter busy-spins <c>Application.DoEvents()</c> for ~1&#160;s
-        /// (Gemini FlatPanel Lite, iOptron, …). Left inline that freezes the GUI (Not Responding), and
-        /// on a host with no message pump it can crash the process. Offloading the call moves that
-        /// blocking prefix off the render loop. The fake devices connect instantly precisely because
-        /// they have no such blocking prefix.
-        /// </summary>
-        private static Task RunDeviceOpOffRenderThreadAsync(Func<Task> deviceOp, CancellationToken ct)
-            => Task.Run(deviceOp, ct);
 
         /// <summary>
         /// Applies site coordinates from a transform to the planner state.
@@ -711,18 +703,6 @@ namespace TianWen.UI.Abstractions
             }
         }
 
-        /// <summary>
-        /// The UTC instant the sky map is currently displaying: the base planning date (or the
-        /// live clock when no date is pinned) PLUS the sky-map time-scrub offset. EVERY sky-map
-        /// handler that computes alt/az or hit-tests an ephemeris position must resolve the
-        /// viewing time through here so the info panel / selection matches what the renderer
-        /// actually drew -- e.g. a scrubbed Zenith reticle must still report Alt 90 deg, not the
-        /// un-scrubbed altitude. The fixed-point / mount info handlers used to omit TimeOffset and
-        /// drifted from the click-select path once the map was scrubbed.
-        /// </summary>
-        private DateTimeOffset SkyMapViewingUtc()
-            => (_plannerState.PlanningDate?.ToUniversalTime() ?? _timeProvider.GetUtcNow()) + _skyMapState.TimeOffset;
-
         public AppSignalHandler(
             IServiceProvider sp,
             GuiAppState appState,
@@ -757,1718 +737,21 @@ namespace TianWen.UI.Abstractions
             _logger = sp.GetRequiredService<ILoggerFactory>().CreateLogger(nameof(AppSignalHandler));
             _timeProvider = sp.GetRequiredService<ITimeProvider>();
 
-            var logger = _logger;
-
-            // ---------------------------------------------------------------
-            // Wire planner search input callbacks
-            // ---------------------------------------------------------------
-            string[]? autoCompleteCache = null;
-
-            plannerState.SearchInput.OnCommit = text =>
-            {
-                plannerState.Suggestions.Clear();
-                plannerState.SuggestionIndex = -1;
-                plannerState.LastSuggestionQuery = "";
-
-                if (appState.ActiveProfile is not null && text.Length > 0)
-                {
-                    var transform = TransformFactory.FromProfile(appState.ActiveProfile, _timeProvider, out _);
-                    if (transform is not null)
-                    {
-                        var db = sp.GetRequiredService<TianWen.Lib.Astrometry.Catalogs.ICelestialObjectDB>();
-                        var resultIdx = PlannerActions.SearchTargets(plannerState, db, transform, text, plannerState.Comets);
-                        if (resultIdx >= 0)
-                        {
-                            plannerState.SelectedTargetIndex = resultIdx;
-                            OnPlannerEnsureVisible?.Invoke(resultIdx);
-                        }
-                    }
-                }
-                return Task.CompletedTask;
-            };
-
-            plannerState.SearchInput.OnCancel = () =>
-            {
-                plannerState.SearchInput.Clear();
-                plannerState.SearchResults = [];
-                plannerState.Suggestions.Clear();
-                plannerState.SuggestionIndex = -1;
-                plannerState.LastSuggestionQuery = "";
-                bus.Post(new DeactivateTextInputSignal());
-                plannerState.NeedsRedraw = true;
-            };
-
-            plannerState.SearchInput.OnTextChanged = text =>
-            {
-                if (autoCompleteCache is not null)
-                {
-                    PlannerActions.UpdateSuggestions(plannerState, autoCompleteCache, text);
-                }
-            };
-
-            // Autocomplete navigation: Up/Down/Return/Escape when suggestions are visible
-            plannerState.SearchInput.OnKeyOverride = key =>
-            {
-                if (plannerState.Suggestions.Count == 0)
-                {
-                    return false;
-                }
-
-                switch (key)
-                {
-                    case TextInputKey.Backspace or TextInputKey.Delete:
-                        return false; // Let the text input handle it, OnTextChanged will update suggestions
-
-                    case TextInputKey.Enter when plannerState.SuggestionIndex >= 0:
-                        CommitSuggestion(plannerState.Suggestions[plannerState.SuggestionIndex]);
-                        return true;
-
-                    case TextInputKey.Escape:
-                        plannerState.Suggestions.Clear();
-                        plannerState.SuggestionIndex = -1;
-                        plannerState.LastSuggestionQuery = "";
-                        appState.NeedsRedraw = true;
-                        return true;
-
-                    default:
-                        return false;
-                }
-            };
-
-            // ---------------------------------------------------------------
-            // Wire sky-map F3 search
-            // ---------------------------------------------------------------
-            var skySearch = skyMapState.Search;
-
-            skySearch.SearchInput.OnTextChanged = _ =>
-            {
-                var db = sp.GetRequiredService<ICelestialObjectDB>();
-                SkyMapSearchActions.FilterResults(skySearch, db);
-                skyMapState.NeedsRedraw = true;
-                appState.NeedsRedraw = true;
-            };
-
-            skySearch.SearchInput.OnCommit = _ =>
-            {
-                bus.Post(new SkyMapSearchCommitSignal());
-                return Task.CompletedTask;
-            };
-
-            skySearch.SearchInput.OnCancel = () => bus.Post(new CloseSkyMapSearchSignal());
-
-            bus.Subscribe<OpenSkyMapSearchSignal>(_ =>
-            {
-                var db = sp.GetRequiredService<ICelestialObjectDB>();
-                SkyMapSearchActions.OpenSearch(skySearch, db, plannerState.Comets);
-                bus.Post(new ActivateTextInputSignal(skySearch.SearchInput));
-                skyMapState.NeedsRedraw = true;
-                appState.NeedsRedraw = true;
-            });
-
-            bus.Subscribe<CloseSkyMapSearchSignal>(_ =>
-            {
-                SkyMapSearchActions.CloseSearch(skySearch);
-                bus.Post(new DeactivateTextInputSignal());
-                skyMapState.NeedsRedraw = true;
-                appState.NeedsRedraw = true;
-            });
-
-            bus.Subscribe<SkyMapSearchCommitSignal>(_ =>
-            {
-                var db = sp.GetRequiredService<ICelestialObjectDB>();
-                // Include the sky-map scrub offset so a planet commit resolves the SAME live position
-                // the map is showing (and reads the render's planet cache without thrashing it).
-                var viewingUtc = SkyMapViewingUtc();
-                var site = SiteContext.Create(plannerState.SiteLatitude, plannerState.SiteLongitude, viewingUtc);
-                SkyMapSearchActions.CommitResult(
-                    skySearch, skyMapState, db,
-                    plannerState.SiteLatitude, plannerState.SiteLongitude,
-                    viewingUtc, site, plannerState.Comets);
-                bus.Post(new DeactivateTextInputSignal());
-                skyMapState.NeedsRedraw = true;
-                appState.NeedsRedraw = true;
-            });
-
-            bus.Subscribe<ViewInPlannerSignal>(sig =>
-            {
-                // Ensure the target is scored + profiled so the planner list has
-                // something to select. Reuses the same CommitSuggestion path the
-                // planner search uses, so alias population and altitude profile
-                // are consistent across the two entry points.
-                var target = new Target(sig.RA, sig.Dec, sig.Name, sig.Index);
-                if (appState.ActiveProfile is { } prof)
-                {
-                    var transform = TransformFactory.FromProfile(prof, _timeProvider, out _);
-                    if (transform is not null && !plannerState.ScoredTargets.ContainsKey(target))
-                    {
-                        var db = sp.GetRequiredService<ICelestialObjectDB>();
-                        PlannerActions.CommitSuggestion(plannerState, db, transform, sig.Name, plannerState.Comets);
-                    }
-                }
-
-                // Find the target in the filtered list and scroll it into view.
-                var filtered = PlannerActions.GetFilteredTargets(plannerState);
-                for (var i = 0; i < filtered.Count; i++)
-                {
-                    if (filtered[i].Target == target)
-                    {
-                        plannerState.SelectedTargetIndex = i;
-                        OnPlannerEnsureVisible?.Invoke(i);
-                        break;
-                    }
-                }
-
-                appState.ActiveTab = GuiTab.Planner;
-                appState.NeedsRedraw = true;
-                plannerState.NeedsRedraw = true;
-            });
-
-            bus.Subscribe<SkyMapPinObjectSignal>(sig =>
-            {
-                var target = new Target(sig.RA, sig.Dec, sig.Name, sig.Index);
-                var transform = appState.ActiveProfile is { } prof
-                    ? TransformFactory.FromProfile(prof, _timeProvider, out _)
-                    : null;
-                var db = sp.GetRequiredService<ICelestialObjectDB>();
-                PlannerActions.TogglePinFromExternal(
-                    plannerState, db, transform, target, sig.ObjectType);
-                bus.Post(new SavePlannerSessionSignal());
-                skyMapState.NeedsRedraw = true;
-                appState.NeedsRedraw = true;
-            });
-
-            bus.Subscribe<SkyMapSlewToObjectSignal>(sig =>
-            {
-                if (!EnsureSessionIdle("Cannot slew manually while a session is running")) return;
-                if (appState.ActiveProfile is not { Data: { } pdata } profile
-                    || pdata.Mount is not { Scheme: not "none" } mountUri)
-                {
-                    Notify(NotificationSeverity.Warning, "No mount configured in the active profile");
-                    return;
-                }
-                if (appState.DeviceHub is not { } hub
-                    || !hub.TryGetConnectedDriver<IMountDriver>(mountUri, out var mount)
-                    || mount is null)
-                {
-                    Notify(NotificationSeverity.Warning, "Mount is not connected \u2014 connect it from the Equipment tab first");
-                    return;
-                }
-
-                // Two-click confirmation for Sun slew. First click arms, second click
-                // within the window proceeds. The arm/confirm state machine lives on
-                // GuiAppState; this routes.
-                if (sig.Index == CatalogIndex.Sol
-                    && appState.GateSunSlew(CatalogIndex.Sol, _timeProvider.GetUtcNow(), TimeSpan.FromSeconds(5))
-                        == GuiAppState.SunSlewGate.Armed)
-                {
-                    appState.StatusMessage =
-                        "\u26A0 SUN \u2014 click Goto again within 5s to confirm. Verify a solar filter is installed.";
-                    skyMapState.NeedsRedraw = true;
-                    appState.NeedsRedraw = true;
-                    return;
-                }
-
-                var minAlt = System.Math.Max((int)plannerState.MinHeightAboveHorizon, 1);
-                var capturedMount = mount;
-                var capturedSig = sig;
-                var capturedProfile = profile;
-                tracker.Run(async () =>
-                {
-                    try
-                    {
-                        var (post, msg) = await MountActions.SlewToJ2000Async(
-                            capturedMount, capturedSig.Name, capturedSig.RA, capturedSig.Dec, capturedSig.Index,
-                            profile: capturedProfile, timeProvider: _timeProvider,
-                            minAboveHorizonDegrees: minAlt, logger: logger,
-                            cancellationToken: cts.Token);
-                        var severity = post == SlewPostCondition.Slewing
-                            ? NotificationSeverity.Info
-                            : NotificationSeverity.Warning;
-                        Notify(severity, msg);
-                        skyMapState.NeedsRedraw = true;
-
-                        // Follow up with a "Reached <name>" / "Slew timed out" notification when
-                        // the mount actually stops slewing, so the status bar isn't permanently
-                        // stuck on the kick-off "Slewing to ..." message.
-                        if (post == SlewPostCondition.Slewing)
-                        {
-                            // Surface the slew destination on the sky map (marker + ETA, the
-                            // latter estimated in the render path from the polled reticle).
-                            // The signal carries J2000 catalog coords, matching the overlay frame.
-                            skyMapState.ActiveSlewTarget = new SlewTargetInfo(
-                                capturedSig.Name, capturedSig.RA, capturedSig.Dec);
-                            skyMapState.SlewEtaSeconds = double.NaN;
-                            // Kick the mount poll so the reticle picks up IsSlewing (and the
-                            // fast slew cadence) this frame instead of up to a steady interval later.
-                            RequestPreviewMountRefresh();
-                            var (completion, completionMsg) = await MountActions.AwaitSlewCompletionAsync(
-                                capturedMount, capturedSig.Name, _timeProvider,
-                                logger: logger, cancellationToken: cts.Token);
-                            var completionSeverity = completion == MountActions.SlewCompletion.Reached
-                                ? NotificationSeverity.Info
-                                : NotificationSeverity.Warning;
-                            Notify(completionSeverity, completionMsg);
-                        }
-                    }
-                    catch (OperationCanceledException oce)
-                    {
-                        // Shutdown / explicit cancel — no notification needed, but log so a
-                        // mid-slew abort is still traceable in the file logger.
-                        logger.LogDebug(oce, "Slew to {Name} cancelled", capturedSig.Name);
-                    }
-                    catch (Exception ex)
-                    {
-                        logger.LogWarning(ex, "Slew to {Name} failed", capturedSig.Name);
-                        Notify(NotificationSeverity.Error, $"Slew failed: {ex.Message}");
-                    }
-                    finally
-                    {
-                        // Slew finished (reached / timed out / cancelled / failed): drop the
-                        // destination marker so it doesn't linger after the mount settles.
-                        skyMapState.ActiveSlewTarget = null;
-                        skyMapState.SlewEtaSeconds = double.NaN;
-                        skyMapState.NeedsRedraw = true;
-                        appState.NeedsRedraw = true;
-                    }
-                }, $"Goto {sig.Name}");
-            });
-
-            bus.Subscribe<SkyMapClickSelectSignal>(sig =>
-            {
-                var rect = skyMapState.LastContentRect;
-                if (rect.Width <= 0 || rect.Height <= 0) return;
-
-                var db = sp.GetRequiredService<ICelestialObjectDB>();
-                var ppr = SkyMapProjection.PixelsPerRadian(rect.Height, skyMapState.FieldOfViewDeg);
-                var cx = rect.X + rect.Width * 0.5f;
-                var cy = rect.Y + rect.Height * 0.5f;
-                // Match the render's viewing instant: base date/now PLUS the sky-map scrub offset
-                // (State.TimeOffset), via SkyMapViewingUtc. Planet positions are ephemeris-computed
-                // and move with time, so hit-testing them (and the panel's alt/az) must use the same
-                // instant the renderer drew -- otherwise a scrubbed planet dot is unclickable.
-                var viewingUtc = SkyMapViewingUtc();
-                var site = SiteContext.Create(plannerState.SiteLatitude, plannerState.SiteLongitude, viewingUtc);
-
-                SkyMapSearchActions.SelectObjectByClick(
-                    skySearch, skyMapState, db,
-                    plannerState.SiteLatitude, plannerState.SiteLongitude,
-                    viewingUtc, site,
-                    sig.ScreenX, sig.ScreenY,
-                    skyMapState.CurrentViewMatrix, ppr, cx, cy,
-                    preferPointSource: (sig.Modifiers & InputModifier.Ctrl) != 0,
-                    pinnedCatalogIndices: PlannerActions.GetPinnedCatalogIndices(plannerState.Proposals),
-                    comets: plannerState.Comets);
-
-                skyMapState.NeedsRedraw = true;
-                appState.NeedsRedraw = true;
-            });
-
-            bus.Subscribe<SkyMapSetViewSignal>(sig =>
-            {
-                // Route-only: the actions helper owns the centre + FOV-clamp + toggle logic.
-                if (SkyMapViewActions.SetView(skyMapState,
-                        sig.CenterRaHours, sig.CenterDecDeg, sig.FieldOfViewDeg,
-                        sig.ShowObjectOverlay, sig.ShowDarkNebulae))
-                {
-                    appState.NeedsRedraw = true;
-                }
-            });
-
-            bus.Subscribe<SkyMapShowFixedPointInfoSignal>(sig =>
-            {
-                // Must include the sky-map scrub offset (via SkyMapViewingUtc): the Zenith/NCP/SCP
-                // reticles are placed by the renderer at the scrubbed LST, so the panel's alt/az
-                // has to be computed at the SAME instant or a scrubbed Zenith reads e.g. +9.9 deg
-                // instead of +90 deg.
-                var viewingUtc = SkyMapViewingUtc();
-                var site = SiteContext.Create(plannerState.SiteLatitude, plannerState.SiteLongitude, viewingUtc);
-                skySearch.InfoPanel = SkyMapInfoPanelData.FromPosition(
-                    sig.Name, sig.RaHours, sig.DecDeg,
-                    plannerState.SiteLatitude, plannerState.SiteLongitude,
-                    viewingUtc, site) with { FixedPoint = sig.FixedPoint };
-                skyMapState.NeedsRedraw = true;
-                appState.NeedsRedraw = true;
-            });
-
-            bus.Subscribe<SkyMapShowMountInfoSignal>(sig =>
-            {
-                // Same scrub-consistency requirement as the fixed-point handler: the mount reticle
-                // is projected at the scrubbed LST, so its reported alt/az must be too.
-                var viewingUtc = SkyMapViewingUtc();
-                var site = SiteContext.Create(plannerState.SiteLatitude, plannerState.SiteLongitude, viewingUtc);
-                skySearch.InfoPanel = SkyMapInfoPanelData.FromMount(
-                    sig.Name, sig.RaHours, sig.DecDeg,
-                    plannerState.SiteLatitude, plannerState.SiteLongitude,
-                    viewingUtc, site);
-                skyMapState.NeedsRedraw = true;
-                appState.NeedsRedraw = true;
-            });
-
-            bus.Subscribe<SkyMapSolveSyncSignal>(sig =>
-            {
-                if (!EnsureSessionIdle("Cannot solve & sync while a session is running")) return;
-                // Re-entrancy guard: ignore a second click while a solve is already in
-                // flight (the button also shows "Solving ..." and drops its handler, but
-                // a queued signal could still arrive). UI-thread-only read here.
-                if (skyMapState.SolveSyncInProgress)
-                {
-                    return;
-                }
-                if (appState.ActiveProfile is not { Data: { } pdata } profile
-                    || pdata.Mount is not { Scheme: not "none" } mountUri)
-                {
-                    Notify(NotificationSeverity.Warning, "No mount configured in the active profile");
-                    return;
-                }
-                if (appState.DeviceHub is not { } hub
-                    || !hub.TryGetConnectedDriver<IMountDriver>(mountUri, out var mount)
-                    || mount is null)
-                {
-                    Notify(NotificationSeverity.Warning, "Mount is not connected \u2014 connect it from the Equipment tab first");
-                    return;
-                }
-                if (pdata.OTAs is not { Length: > 0 } otas || sig.OtaIndex >= otas.Length)
-                {
-                    Notify(NotificationSeverity.Warning, "No OTA configured in the active profile");
-                    return;
-                }
-                var ota = otas[sig.OtaIndex];
-                if (!TryGetConnected<ICameraDriver>(hub, ota.Camera, "Camera", out var camera)) return;
-
-                // Optional per-OTA devices for FITS denorm stamping (same as TakePreview);
-                // the mount is resolved separately above, so discard it here.
-                var (ssFocuser, ssFilterWheel, _) = EquipmentActions.ResolveOtaCaptureDevices(hub, pdata, sig.OtaIndex);
-
-                // Mirror the preview-capture progress UI while the solve frame exposes.
-                // ExposureSeconds is now trustworthy even for the button's parameterless
-                // `new SkyMapSolveSyncSignal()` thanks to its explicit parameterless ctor.
-                if (sig.OtaIndex < liveSessionState.PreviewCapturing.Length)
-                {
-                    liveSessionState.PreviewCapturing[sig.OtaIndex] = true;
-                    liveSessionState.PreviewCaptureStart[sig.OtaIndex] = _timeProvider.GetUtcNow();
-                    liveSessionState.PreviewExposureDuration[sig.OtaIndex] = TimeSpan.FromSeconds(sig.ExposureSeconds);
-                }
-                appState.StatusMessage = "Solve & sync\u2026";
-                skyMapState.SolveSyncInProgress = true; // drives the "Solving ..." button label
-                appState.NeedsRedraw = true;
-
-                var capturedSig = sig;
-                var capturedMount = mount;
-                var capturedCamera = camera;
-                var capturedProfile = profile;
-                var capturedOta = ota;
-                tracker.Run(async () =>
-                {
-                    try
-                    {
-                        var outcome = await MountActions.SolveAndSyncAsync(
-                            capturedMount, capturedCamera,
-                            capturedOta.Name, capturedOta.FocalLength, capturedOta.Aperture,
-                            ssFocuser, ssFilterWheel,
-                            sp.GetRequiredService<ICelestialObjectDB>(),
-                            sp.GetRequiredService<IPlateSolverFactory>(),
-                            capturedProfile, _timeProvider,
-                            TimeSpan.FromSeconds(capturedSig.ExposureSeconds),
-                            capturedSig.Gain is { } g ? (short)g : null,
-                            capturedSig.Binning,
-                            logger, cts.Token);
-
-                        // Stash the solve frame into the preview slot (ownership transfer from
-                        // the outcome; release the previous occupant so its ChannelBuffer drops).
-                        if (outcome.CapturedImage is { } image
-                            && capturedSig.OtaIndex < liveSessionState.LastCapturedImages.Length)
-                        {
-                            liveSessionState.LastCapturedImages[capturedSig.OtaIndex]?.Release();
-                            liveSessionState.LastCapturedImages[capturedSig.OtaIndex] = image;
-                        }
-                        else
-                        {
-                            outcome.CapturedImage?.Release();
-                        }
-                        if (outcome.SolveResult is { } solveResult)
-                        {
-                            liveSessionState.PreviewPlateSolveResult = solveResult;
-                        }
-
-                        var severity = outcome.Result == MountActions.SolveSyncResult.Synced
-                            ? NotificationSeverity.Info
-                            : NotificationSeverity.Warning;
-                        Notify(severity, outcome.StatusMessage);
-                    }
-                    catch (OperationCanceledException oce)
-                    {
-                        // Shutdown / explicit cancel - log so a mid-solve abort stays traceable.
-                        logger.LogDebug(oce, "Solve & sync cancelled");
-                    }
-                    catch (Exception ex)
-                    {
-                        logger.LogWarning(ex, "Solve & sync failed");
-                        Notify(NotificationSeverity.Error, $"Solve & sync failed: {ex.Message}");
-                    }
-                    finally
-                    {
-                        if (capturedSig.OtaIndex < liveSessionState.PreviewCapturing.Length)
-                        {
-                            liveSessionState.PreviewCapturing[capturedSig.OtaIndex] = false;
-                        }
-                        skyMapState.SolveSyncInProgress = false; // re-enable the Solve & Sync button
-                        // A sync doesn't change slew/track state, so the cadence ramp won't
-                        // notice it - force a poll so the reticle jumps to the synced pointing.
-                        RequestPreviewMountRefresh();
-                        skyMapState.NeedsRedraw = true;
-                        liveSessionState.NeedsRedraw = true;
-                        appState.NeedsRedraw = true;
-                    }
-                }, "SolveAndSync");
-            });
-
-            // ---------------------------------------------------------------
-            // Wire equipment text input callbacks
-            // ---------------------------------------------------------------
-
-            eqState.ProfileNameInput.OnCommit = async text =>
-            {
-                if (text.Length > 0)
-                {
-                    var profile = await EquipmentActions.CreateProfileAsync(text, external, cts.Token);
-                    appState.ActiveProfile = profile;
-                    eqState.IsCreatingProfile = false;
-                    eqState.ProfileNameInput.Clear();
-                    bus.Post(new DeactivateTextInputSignal());
-                    plannerState.NeedsRecompute = true;
-                    appState.NeedsRedraw = true;
-                }
-            };
-
-            eqState.ProfileNameInput.OnCancel = () =>
-            {
-                eqState.IsCreatingProfile = false;
-                eqState.ProfileNameInput.Clear();
-            };
-
-            // Site inputs share a commit: save site on Enter from any of the three fields.
-            // Parse/validate + the mount push live in EquipmentActions; this routes.
-            Func<Task> saveSite = async () =>
-            {
-                if (appState.ActiveProfile is not { } siteProfile)
-                {
-                    return;
-                }
-
-                if (!EquipmentActions.TryParseSite(
-                        eqState.LatitudeInput.Text, eqState.LongitudeInput.Text, eqState.ElevationInput.Text,
-                        out var sLat, out var sLon, out var sElev))
-                {
-                    Notify(NotificationSeverity.Warning, "Invalid coordinates (lat: -90..90, lon: -180..180)");
-                    return;
-                }
-
-                var sData = siteProfile.Data ?? ProfileData.Empty;
-                var newSiteData = EquipmentActions.SetSite(sData, sLat, sLon, sElev);
-                var updatedSite = siteProfile.WithData(newSiteData);
-                // Update UI immediately, save in background
-                appState.ActiveProfile = updatedSite;
-                eqState.IsEditingSite = false;
-                bus.Post(new DeactivateTextInputSignal());
-                plannerState.SiteLatitude = sLat;
-                plannerState.SiteLongitude = sLon;
-                plannerState.NeedsRecompute = true;
-                appState.NeedsRedraw = true;
-
-                // If the catalog was blocked on a missing site, load it now.
-                if (plannerState.ObjectDb is null
-                    && TransformFactory.FromProfile(updatedSite, _timeProvider, out _) is { } siteTransform)
-                {
-                    _tracker.Run(() => InitializePlannerAsync(siteTransform, cts.Token),
-                        "Load catalog after site edit");
-                }
-
-                await EquipmentActions.PushSiteToMountIfProfileWinsAsync(
-                    appState.DeviceHub, newSiteData, sLat, sLon, sElev, logger, cts.Token);
-                await updatedSite.SaveAsync(external, cts.Token);
-            };
-
-            Action cancelSite = () =>
-            {
-                eqState.IsEditingSite = false;
-                eqState.LatitudeInput.Deactivate();
-                eqState.LongitudeInput.Deactivate();
-                eqState.ElevationInput.Deactivate();
-                appState.ActiveTextInput = null;
-            };
-
-            eqState.LatitudeInput.OnCommit = _ => saveSite();
-            eqState.LongitudeInput.OnCommit = _ => saveSite();
-            eqState.ElevationInput.OnCommit = _ => saveSite();
-            eqState.LatitudeInput.OnCancel = cancelSite;
-            eqState.LongitudeInput.OnCancel = cancelSite;
-            eqState.ElevationInput.OnCancel = cancelSite;
-
-            // Guide scope focal length — commit on Enter
-            eqState.GuiderFocalLengthInput.OnCommit = async text =>
-            {
-                if (appState.ActiveProfile is { } profile && profile.Data is { } pd)
-                {
-                    int? guiderFl = int.TryParse(text, out var fl) && fl > 0 ? fl : null;
-                    var updated = profile.WithData(pd with { GuiderFocalLength = guiderFl });
-                    appState.ActiveProfile = updated;
-                    appState.NeedsRedraw = true;
-                    await updated.SaveAsync(external, cts.Token);
-                }
-            };
-
-            // OTA name / focal length / aperture — commit on Enter saves the OTA edit
-            Task saveOta(string _)
-            {
-                if (appState.ActiveProfile is { Data: { } editData } && eqState.EditingOtaIndex >= 0)
-                {
-                    var otaIdx = eqState.EditingOtaIndex;
-                    var newName = eqState.OtaNameInput.Text is { Length: > 0 } n ? n : null;
-                    int? newFl = int.TryParse(eqState.FocalLengthInput.Text, out var fl) && fl > 0 ? fl : null;
-                    int? newAp = int.TryParse(eqState.ApertureInput.Text, out var ap) ? ap : null;
-                    var newData = EquipmentActions.UpdateOTA(editData, otaIdx, name: newName, focalLength: newFl, aperture: newAp);
-                    bus.Post(new UpdateProfileSignal(newData));
-                    eqState.StopEditingOta();
-                }
-                return Task.CompletedTask;
-            }
-            eqState.OtaNameInput.OnCommit = saveOta;
-            eqState.FocalLengthInput.OnCommit = saveOta;
-            eqState.ApertureInput.OnCommit = saveOta;
-
-            // Device string settings (API keys, ports, etc.) — commit on Enter saves the setting
-            eqState.StringSettingInput.OnCommit = async _ =>
-            {
-                if (eqState.EditingStringSettingKey is not { } key || eqState.EditingDeviceUri is not { } editUri)
-                {
-                    return;
-                }
-
-                var value = eqState.StringSettingInput.Text;
-                eqState.EditingStringSettingKey = null;
-
-                // The masked-secret-vs-URI-param decision (credential-store write for secrets,
-                // query-param URI for the rest) lives in EquipmentActions; this routes.
-                var commit = EquipmentActions.CommitDeviceSetting(
-                    editUri, key, value, sp.GetRequiredService<ICredentialStore>());
-                if (commit.Kind == EquipmentActions.DeviceSettingCommitKind.StoredSecret)
-                {
-                    // The URI/profile is unchanged, so the refetch-on-weather-URI-change path won't
-                    // fire; re-fetch here now that the key may have become available.
-                    if (commit.IsWeatherSecret)
-                    {
-                        await FetchWeatherForecastAsync(cts.Token);
-                    }
-                    appState.NeedsRedraw = true;
-                    return;
-                }
-
-                // Non-secret: keep as a query param on the device URI (existing behaviour).
-                eqState.EditingDeviceUri = commit.NewUri;
-                if (appState.ActiveProfile is { Data: { } data } && commit.NewUri is { } newUri
-                    && eqState.SavedDeviceSettingsUri is { } savedUri)
-                {
-                    var newData = EquipmentActions.UpdateDeviceUri(data, savedUri, newUri);
-                    bus.Post(new UpdateProfileSignal(newData));
-                    eqState.BeginEditingDeviceSettings(newUri);
-                }
-            };
-
-            // ---------------------------------------------------------------
-            // Equipment action signal subscriptions (DI-dependent handlers)
-            // ---------------------------------------------------------------
-
-            bus.Subscribe<DiscoverDevicesSignal>(async sig =>
-            {
-                if (eqState.IsDiscovering) return;
-
-                eqState.IsDiscovering = true;
-                appState.StatusMessage = sig.IncludeFake ? "Discovering devices (+ fake)..." : "Discovering devices...";
-                appState.NeedsRedraw = true;
-                try
-                {
-                    var dm = sp.GetRequiredService<IDeviceDiscovery>();
-                    await dm.CheckSupportAsync(cts.Token);
-                    await dm.DiscoverAsync(cts.Token);
-                    eqState.DiscoveredDevices = [.. dm.RegisteredDeviceTypes
-                        .Where(t => t is not DeviceType.Profile and not DeviceType.None)
-                        .SelectMany(dm.RegisteredDevices)
-                        .Where(d => sig.IncludeFake || d is not TianWen.Lib.Devices.Fake.FakeDevice)
-                        .OrderBy(d => d.DeviceType).ThenBy(d => d.DisplayName)];
-
-                    // Route post-discovery profile reconciliation to EquipmentActions,
-                    // then reflect any active-profile update back into UI state.
-                    var reconciledProfiles = await EquipmentActions.ReconcileAllProfilesAsync(dm, external, cts.Token);
-                    foreach (var (original, updated) in reconciledProfiles)
-                    {
-                        if (appState.ActiveProfile?.ProfileId == original.ProfileId)
-                        {
-                            appState.ActiveProfile = updated;
-                        }
-
-                        // Log exactly what URI moved so site / gain / filter clobbers are
-                        // visible in the log instead of silently drifting.
-                        var diffs = EquipmentActions.DiffProfileData(original.Data!.Value, updated.Data!.Value);
-                        foreach (var (field, before, after) in diffs)
-                        {
-                            logger.LogInformation(
-                                "Reconcile {Profile} {Field}: {Before} -> {After}",
-                                original.DisplayName, field, before, after);
-                        }
-                    }
-                    if (reconciledProfiles.Count > 0)
-                    {
-                        logger.LogInformation("Post-discovery reconcile: updated {Count} profile(s)", reconciledProfiles.Count);
-                    }
-                }
-                catch (Exception ex)
-                {
-                    logger.LogWarning(ex, "Device discovery failed");
-                    Notify(NotificationSeverity.Error, "Discovery failed");
-                }
-                finally
-                {
-                    eqState.IsDiscovering = false;
-                    if (eqState.DiscoveredDevices.Count > 0)
-                    {
-                        Notify(NotificationSeverity.Info, $"Found {eqState.DiscoveredDevices.Count} devices");
-                    }
-                    appState.NeedsRedraw = true;
-                }
-            });
-
-            bus.Subscribe<AddOtaSignal>(async _ =>
-            {
-                if (appState.ActiveProfile is not { } p) return;
-
-                var data = p.Data ?? ProfileData.Empty;
-                var newOta = new OTAData(
-                    Name: $"Telescope #{data.OTAs.Length}",
-                    FocalLength: 1000,
-                    Camera: NoneDevice.Instance.DeviceUri,
-                    Cover: null, Focuser: null, FilterWheel: null,
-                    PreferOutwardFocus: null, OutwardIsPositive: null,
-                    Aperture: null, OpticalDesign: OpticalDesign.Unknown);
-                var updated = p.WithData(EquipmentActions.AddOTA(data, newOta));
-                appState.ActiveProfile = updated;
-                appState.NeedsRedraw = true;
-                await updated.SaveAsync(external, cts.Token);
-            });
-
-            bus.Subscribe<EditSiteSignal>(_ =>
-            {
-                eqState.IsEditingSite = true;
-                if (appState.ActiveProfile?.Data is { } pd)
-                {
-                    var existingSite = EquipmentActions.GetSiteFromProfile(pd);
-                    if (existingSite.HasValue)
-                    {
-                        eqState.LatitudeInput.Text = existingSite.Value.Lat.ToString(System.Globalization.CultureInfo.InvariantCulture);
-                        eqState.LatitudeInput.CursorPos = eqState.LatitudeInput.Text.Length;
-                        eqState.LongitudeInput.Text = existingSite.Value.Lon.ToString(System.Globalization.CultureInfo.InvariantCulture);
-                        eqState.LongitudeInput.CursorPos = eqState.LongitudeInput.Text.Length;
-                        eqState.ElevationInput.Text = existingSite.Value.Elev?.ToString(System.Globalization.CultureInfo.InvariantCulture) ?? "";
-                        eqState.ElevationInput.CursorPos = eqState.ElevationInput.Text.Length;
-                    }
-                }
-                bus.Post(new ActivateTextInputSignal(eqState.LatitudeInput));
-            });
-
-            bus.Subscribe<CreateProfileSignal>(_ =>
-            {
-                if (!eqState.IsCreatingProfile)
-                {
-                    eqState.IsCreatingProfile = true;
-                    bus.Post(new ActivateTextInputSignal(eqState.ProfileNameInput));
-                }
-            });
-
-            bus.Subscribe<AssignDeviceSignal>(async sig =>
-            {
-                var deviceIndex = sig.DeviceIndex;
-                if (deviceIndex < 0 || deviceIndex >= eqState.DiscoveredDevices.Count) return;
-
-                if (eqState.ActiveAssignment is { } target && appState.ActiveProfile is { } profile)
-                {
-                    var device = eqState.DiscoveredDevices[deviceIndex];
-
-                    if (device.DeviceType != target.ExpectedDeviceType)
-                    {
-                        Notify(NotificationSeverity.Warning, $"Expected {target.ExpectedDeviceType}, got {device.DeviceType}");
-                        return;
-                    }
-
-                    var data = profile.Data ?? ProfileData.Empty;
-
-                    // Capture the URI previously assigned to THIS slot. If still connected
-                    // via the hub, we'll auto-disconnect it after assignment iff safe
-                    // (cooler off, idle). Cool/busy orphans are left connected and the
-                    // user is told to disconnect manually so warm-up runs.
-                    var prevSlotUri = EquipmentActions.GetAssignedDevice(data, target);
-
-                    data = EquipmentActions.UnassignDevice(data, device.DeviceUri);
-
-                    var newData = EquipmentActions.ApplyAssignment(data, target, device.DeviceType, device.DeviceUri);
-
-                    var updated = profile.WithData(newData);
-                    appState.ActiveProfile = updated;
-                    // Keep the slot active so the user can swap the assigned device by
-                    // clicking another row immediately, without re-clicking the slot.
-                    // Click the slot itself again to deactivate.
-                    appState.NeedsRedraw = true;
-                    await updated.SaveAsync(external, cts.Token);
-
-                    // Fetch weather forecast immediately when a weather device is assigned
-                    if (device.DeviceType is DeviceType.Weather)
-                    {
-                        await FetchWeatherForecastAsync(cts.Token);
-                        appState.NeedsRedraw = true;
-                    }
-
-                    // Auto-disconnect the orphan if it's still connected and safe
-                    // (EquipmentActions.AutoDisconnectOrphanAsync); this lambda only maps
-                    // the outcome to notifications.
-                    if (appState.DeviceHub is { } hub)
-                    {
-                        var (outcome, safety) = await EquipmentActions.AutoDisconnectOrphanAsync(
-                            hub, prevSlotUri, device.DeviceUri, logger, cts.Token);
-                        switch (outcome)
-                        {
-                            case EquipmentActions.OrphanDisconnectOutcome.Disconnected:
-                                Notify(NotificationSeverity.Info, $"Previous {target.ExpectedDeviceType} disconnected");
-                                break;
-                            case EquipmentActions.OrphanDisconnectOutcome.LeftConnected:
-                                Notify(NotificationSeverity.Warning, $"Previous {target.ExpectedDeviceType} left connected ({safety}). Click Off on its row to warm up.");
-                                break;
-                        }
-                        if (outcome != EquipmentActions.OrphanDisconnectOutcome.NotApplicable)
-                        {
-                            appState.NeedsRedraw = true;
-                        }
-                    }
-                }
-            });
-
-            bus.Subscribe<AssignManualCoverSignal>(async _ =>
-            {
-                // The manual light panel is not discoverable, so it never appears in the device list.
-                // Assign its canonical URI straight to the active Cover slot (mirrors the URI-based tail
-                // of AssignDeviceSignal). It then flows through the ordinary calibrator flat path.
-                if (eqState.ActiveAssignment is not { } target || appState.ActiveProfile is not { } profile)
-                {
-                    return;
-                }
-                if (target.ExpectedDeviceType != DeviceType.CoverCalibrator)
-                {
-                    Notify(NotificationSeverity.Warning, "Select an OTA's Cover slot first, then add the Manual Light Panel");
-                    return;
-                }
-
-                var manual = new TianWen.Lib.Devices.ManualCoverDevice();
-                var data = profile.Data ?? ProfileData.Empty;
-                var newData = EquipmentActions.ApplyAssignment(data, target, DeviceType.CoverCalibrator, manual.DeviceUri);
-                var updated = profile.WithData(newData);
-                appState.ActiveProfile = updated;
-                appState.NeedsRedraw = true;
-                await updated.SaveAsync(external, cts.Token);
-                Notify(NotificationSeverity.Info, "Manual Light Panel assigned - switch it on before capturing flats");
-            });
-
-            bus.Subscribe<ConnectAllDevicesSignal>(_ =>
-            {
-                if (appState.ActiveProfile?.Data is not { } pdata) return;
-                if (appState.DeviceHub is not { } hub) return;
-
-                // Fan out to per-device ConnectDeviceSignal so each connect goes through
-                // the same in-flight gate, notification, and safety paths as a manual
-                // click. Skip URIs that the hub already considers connected — connecting
-                // an already-connected URI just churns PendingTransitions without effect.
-                foreach (var uri in pdata.AssignedDeviceUris)
-                {
-                    if (hub.IsConnected(uri)) continue;
-                    bus.Post(new ConnectDeviceSignal(uri));
-                }
-            });
-
-            bus.Subscribe<ConnectDeviceSignal>(async sig =>
-            {
-                if (appState.DeviceHub is not { } hub)
-                {
-                    Notify(NotificationSeverity.Warning, "Device hub unavailable");
-                    return;
-                }
-
-                if (!eqState.PendingTransitions.TryAdd(sig.DeviceUri, 0))
-                {
-                    return; // transition already in flight
-                }
-                appState.NeedsRedraw = true;
-
-                try
-                {
-                    // Prefer the resolved device (carries query-param config); fall back
-                    // to a freshly-discovered match by URI equality.
-                    var device = EquipmentActions.ResolveDeviceForConnect(hub, eqState.DiscoveredDevices, sig.DeviceUri);
-
-                    if (device is null)
-                    {
-                        Notify(NotificationSeverity.Warning, "Cannot resolve device URI for connect");
-                        return;
-                    }
-
-                    await RunDeviceOpOffRenderThreadAsync(() => hub.ConnectAsync(device, cts.Token).AsTask(), cts.Token);
-                    Notify(NotificationSeverity.Info, $"Connected: {device.DisplayName}");
-
-                    // Mount connect → reconcile site between mount hardware and profile,
-                    // per SiteTieBreaker. Updates ProfileState + PlannerState; persists
-                    // profile if adopted; pushes to mount if profile wins.
-                    if (device.DeviceType == DeviceType.Mount
-                        && appState.ActiveProfile is { } currentProfile
-                        && currentProfile.Data is { } pdata
-                        && DeviceBase.SameDevice(pdata.Mount, sig.DeviceUri)
-                        && hub.TryGetConnectedDriver<IMountDriver>(sig.DeviceUri, out var mount)
-                        && mount is not null)
-                    {
-                        var outcome = await EquipmentActions.ReconcileSiteOnMountConnectAsync(
-                            pdata, mount, logger, cts.Token);
-                        if (outcome.ProfileChanged)
-                        {
-                            var updated = currentProfile.WithData(outcome.Data);
-                            await updated.SaveAsync(_external, cts.Token);
-                            appState.ActiveProfile = updated;
-                        }
-                        if (outcome.Data.SiteLatitude is { } rlat && outcome.Data.SiteLongitude is { } rlon)
-                        {
-                            plannerState.SiteLatitude = rlat;
-                            plannerState.SiteLongitude = rlon;
-                            plannerState.NeedsRedraw = true;
-
-                            // If the catalog hasn't loaded yet because we previously
-                            // had no site, fire InitializePlannerAsync now.
-                            if (plannerState.ObjectDb is null
-                                && TransformFactory.FromProfile(appState.ActiveProfile!, _timeProvider, out _) is { } rTransform)
-                            {
-                                _tracker.Run(() => InitializePlannerAsync(rTransform, cts.Token),
-                                    "Load catalog after site reconcile");
-                            }
-                        }
-                    }
-
-                    // Camera connect -> capture sensor geometry into the matching OTA so the planner
-                    // can compute the sensor FOV (smart framing groups) offline afterwards. Idempotent:
-                    // CaptureSensorSpecs returns null (no save) once the specs are stored and unchanged.
-                    // ActiveProfile is re-read fresh here; a rare ConnectAll multi-camera race can drop
-                    // one OTA's first capture, which self-heals on the next connect.
-                    if (device.DeviceType == DeviceType.Camera
-                        && appState.ActiveProfile is { } camProfile
-                        && camProfile.Data is { } camData
-                        && hub.TryGetConnectedDriver<ICameraDriver>(sig.DeviceUri, out var cam)
-                        && cam is not null
-                        && EquipmentActions.CaptureSensorSpecs(camData, sig.DeviceUri, cam) is { } capturedData)
-                    {
-                        var updated = camProfile.WithData(capturedData);
-                        await updated.SaveAsync(_external, cts.Token);
-                        appState.ActiveProfile = updated;
-                        // The FOV is now known -> push it to the planner and (re)compute framing groups.
-                        RefreshSensorFovAndFraming();
-                        logger.LogInformation("Captured sensor geometry for {Uri} into profile", sig.DeviceUri);
-                    }
-                }
-                catch (Exception ex)
-                {
-                    logger.LogWarning(ex, "Connect failed for {Uri}", sig.DeviceUri);
-                    Notify(NotificationSeverity.Error, $"Connect failed: {ex.Message}");
-                }
-                finally
-                {
-                    eqState.PendingTransitions.TryRemove(sig.DeviceUri, out _);
-                    appState.NeedsRedraw = true;
-                }
-            });
-
-            bus.Subscribe<DisconnectDeviceSignal>(async sig =>
-            {
-                if (appState.DeviceHub is not { } hub)
-                {
-                    return;
-                }
-
-                // Pre-flight safety check. If the device is a cooled/busy camera, don't
-                // disconnect — set the per-row confirmation state so the UI shows the
-                // [Warm & Off] [Force Off] [Cancel] strip instead of executing.
-                var safety = await EquipmentActions.GetDisconnectSafetyAsync(hub, sig.DeviceUri, cts.Token);
-                if (safety != EquipmentActions.DisconnectSafety.Safe)
-                {
-                    eqState.PendingDisconnectConfirm = sig.DeviceUri;
-                    eqState.PendingDisconnectSafety = safety;
-                    eqState.PendingForceConfirm = null;
-                    appState.NeedsRedraw = true;
-                    return;
-                }
-
-                if (!eqState.PendingTransitions.TryAdd(sig.DeviceUri, 0))
-                {
-                    return;
-                }
-                appState.NeedsRedraw = true;
-
-                try
-                {
-                    await RunDeviceOpOffRenderThreadAsync(() => hub.DisconnectAsync(sig.DeviceUri, cts.Token).AsTask(), cts.Token);
-                    Notify(NotificationSeverity.Info, "Device disconnected");
-                }
-                catch (Exception ex)
-                {
-                    logger.LogWarning(ex, "Disconnect failed for {Uri}", sig.DeviceUri);
-                    Notify(NotificationSeverity.Error, $"Disconnect failed: {ex.Message}");
-                }
-                finally
-                {
-                    eqState.PendingTransitions.TryRemove(sig.DeviceUri, out _);
-                    appState.NeedsRedraw = true;
-                }
-            });
-
-            bus.Subscribe<ForceDisconnectDeviceSignal>(async sig =>
-            {
-                if (appState.DeviceHub is not { } hub)
-                {
-                    return;
-                }
-
-                // Bypass the safety pre-check. Caller already passed two-stage confirmation.
-                eqState.PendingDisconnectConfirm = null;
-                eqState.PendingForceConfirm = null;
-                if (!eqState.PendingTransitions.TryAdd(sig.DeviceUri, 0))
-                {
-                    return;
-                }
-                appState.NeedsRedraw = true;
-
-                try
-                {
-                    await RunDeviceOpOffRenderThreadAsync(() => hub.DisconnectAsync(sig.DeviceUri, cts.Token).AsTask(), cts.Token);
-                    Notify(NotificationSeverity.Info, "Device force-disconnected (no warm-up)");
-                    logger.LogWarning("Force-disconnect of {Uri} (bypassed safety check)", sig.DeviceUri);
-                }
-                catch (Exception ex)
-                {
-                    logger.LogWarning(ex, "Force-disconnect failed for {Uri}", sig.DeviceUri);
-                    Notify(NotificationSeverity.Error, $"Force-disconnect failed: {ex.Message}");
-                }
-                finally
-                {
-                    eqState.PendingTransitions.TryRemove(sig.DeviceUri, out _);
-                    appState.NeedsRedraw = true;
-                }
-            });
-
-            bus.Subscribe<WarmAndDisconnectDeviceSignal>(async sig =>
-            {
-                if (appState.DeviceHub is not { } hub)
-                {
-                    return;
-                }
-
-                eqState.PendingDisconnectConfirm = null;
-                eqState.PendingForceConfirm = null;
-                if (!eqState.PendingTransitions.TryAdd(sig.DeviceUri, 0))
-                {
-                    return;
-                }
-                appState.NeedsRedraw = true;
-
-                try
-                {
-                    await RunDeviceOpOffRenderThreadAsync(() => EquipmentActions.WarmAndDisconnectAsync(hub, sig.DeviceUri, _logger, cts.Token).AsTask(), cts.Token);
-                    Notify(NotificationSeverity.Info, "Camera warmed and disconnected");
-                }
-                catch (OperationCanceledException)
-                {
-                    Notify(NotificationSeverity.Warning, "Warm-up cancelled");
-                }
-                catch (Exception ex)
-                {
-                    logger.LogWarning(ex, "Warm-and-disconnect failed for {Uri}", sig.DeviceUri);
-                    Notify(NotificationSeverity.Error, $"Warm-up failed: {ex.Message}");
-                }
-                finally
-                {
-                    eqState.PendingTransitions.TryRemove(sig.DeviceUri, out _);
-                    appState.NeedsRedraw = true;
-                }
-            });
-
-            bus.Subscribe<SetCoolerSetpointSignal>(async sig =>
-            {
-                if (appState.DeviceHub is not { } hub) return;
-                if (!TryGetConnected<ICameraDriver>(hub, sig.DeviceUri, "Camera", out var camera)) return;
-                try
-                {
-                    await EquipmentActions.SetCoolerSetpointAsync(camera, sig.SetpointC, cts.Token);
-                    Notify(NotificationSeverity.Info, $"Cooling to {sig.SetpointC:F1}\u00b0C");
-                }
-                catch (Exception ex)
-                {
-                    logger.LogWarning(ex, "SetCoolerSetpoint failed for {Uri}", sig.DeviceUri);
-                    Notify(NotificationSeverity.Error, $"Cooler setpoint failed: {ex.Message}");
-                }
-            });
-
-            bus.Subscribe<WarmAndCoolerOffSignal>(async sig =>
-            {
-                if (appState.DeviceHub is not { } hub) return;
-                eqState.PendingCoolerOffConfirm = null;
-                eqState.PendingCoolerOffForceConfirm = null;
-                appState.NeedsRedraw = true;
-
-                try
-                {
-                    await EquipmentActions.WarmAndCoolerOffAsync(hub, sig.DeviceUri, _logger, cts.Token);
-                    Notify(NotificationSeverity.Info, "Camera warmed; cooler off");
-                }
-                catch (OperationCanceledException)
-                {
-                    Notify(NotificationSeverity.Warning, "Warm-up cancelled");
-                }
-                catch (Exception ex)
-                {
-                    logger.LogWarning(ex, "Warm-and-cooler-off failed for {Uri}", sig.DeviceUri);
-                    Notify(NotificationSeverity.Error, $"Warm-up failed: {ex.Message}");
-                }
-                finally { appState.NeedsRedraw = true; }
-            });
-
-            bus.Subscribe<SetCoolerOffSignal>(async sig =>
-            {
-                if (appState.DeviceHub is not { } hub) return;
-                if (!hub.TryGetConnectedDriver<TianWen.Lib.Devices.ICameraDriver>(sig.DeviceUri, out var camera))
-                {
-                    return;
-                }
-                try
-                {
-                    await EquipmentActions.SetCoolerOffAsync(camera, cts.Token);
-                    Notify(NotificationSeverity.Info, "Cooler off");
-                }
-                catch (Exception ex)
-                {
-                    logger.LogWarning(ex, "SetCoolerOff failed for {Uri}", sig.DeviceUri);
-                    Notify(NotificationSeverity.Error, $"Cooler off failed: {ex.Message}");
-                }
-            });
-
-            bus.Subscribe<UpdateProfileSignal>(async sig =>
-            {
-                if (appState.ActiveProfile is { } profile)
-                {
-                    var previousWeather = profile.Data?.Weather;
-                    var updated = profile.WithData(sig.Data);
-                    appState.ActiveProfile = updated;
-                    appState.NeedsRedraw = true;
-                    await updated.SaveAsync(external, cts.Token);
-
-                    // Camera / focuser / filter-wheel assignments may have changed; rebuild the per-OTA
-                    // settings so gain modes (DSLR ISO vs ZWO numeric) and cooling capability come from
-                    // the new device's driver instead of being cached from the old one.
-                    sessionState.InitializeFromProfile(updated, appState.DeviceHub);
-                    sessionState.NeedsRedraw = true;
-
-                    // Refetch weather when the weather device URI changes (e.g. API key entered)
-                    if (sig.Data.Weather != previousWeather)
-                    {
-                        await FetchWeatherForecastAsync(cts.Token);
-                        appState.NeedsRedraw = true;
-                    }
-                }
-            });
-
-            bus.Subscribe<SavePlannerSessionSignal>(async _ =>
-            {
-                if (!plannerState.IsDirty || appState.ActiveProfile is not { } profile)
-                {
-                    return;
-                }
-                plannerState.IsDirty = false;
-                await PlannerPersistence.SaveAsync(plannerState, profile, external, _timeProvider, cts.Token);
-            });
-
-            bus.Subscribe<SaveSessionConfigSignal>(async _ =>
-            {
-                if (!sessionState.IsDirty || appState.ActiveProfile is not { } profile)
-                {
-                    return;
-                }
-                sessionState.IsDirty = false;
-                await SessionPersistence.SaveAsync(sessionState, profile, external, cts.Token);
-            });
-
-            // Wire signal bus into state objects for auto-posting on dirty
-            plannerState.Bus = bus;
-            sessionState.Bus = bus;
-
-            // Refresh per-OTA camera capabilities when a driver connects or disconnects via the hub —
-            // gain modes / cooling info may only become known after the driver is actually instantiated.
-            if (appState.DeviceHub is { } hub)
-            {
-                hub.DeviceStateChanged += (_, _) =>
-                {
-                    if (appState.ActiveProfile is { } profile)
-                    {
-                        sessionState.InitializeFromProfile(profile, hub);
-                        sessionState.NeedsRedraw = true;
-                        appState.NeedsRedraw = true;
-                    }
-                };
-            }
-
-            // ---------------------------------------------------------------
-            // Schedule building (shared between planner preview and session start)
-            // ---------------------------------------------------------------
-            bus.Subscribe<BuildScheduleSignal>(signal =>
-            {
-                if (appState.ActiveProfile is not { } profile) return;
-                var transform = TransformFactory.FromProfile(profile, _timeProvider, out _);
-                if (transform is null) return;
-
-                var profileData = profile.Data ?? ProfileData.Empty;
-                var (filters, design) = GetFirstOtaFilterConfig(profileData);
-                // Use the SAME effective defaults the session-start path uses so the planner
-                // PREVIEW matches what the session will actually capture (it previously hardcoded
-                // gain=120/offset=10/sub=120s here, diverging from both the config and the
-                // f-ratio exposure shown on the target rows). Gain/offset stay null so the per-OTA
-                // camera settings drive them, exactly as StartSession does.
-                PlannerActions.BuildSchedule(plannerState, sessionState, transform,
-                    defaultGain: null, defaultOffset: null,
-                    defaultSubExposure: SessionContent.EffectiveDefaultSubExposure(sessionState),
-                    defaultObservationTime: TimeSpan.FromMinutes(60),
-                    availableFilters: filters,
-                    opticalDesign: design);
-            });
-
-            // ---------------------------------------------------------------
-            // Live session signals
-            // ---------------------------------------------------------------
-
-            bus.Subscribe<StartSessionSignal>(async _ =>
-            {
-                if (!EnsureSessionIdle("Session already running")) return;
-
-                if (appState.ActiveProfile is not { } profile)
-                {
-                    Notify(NotificationSeverity.Warning, "No profile selected");
-                    return;
-                }
-
-                if (plannerState.Proposals is not { Length: > 0 })
-                {
-                    Notify(NotificationSeverity.Warning, "No targets \u2014 pin targets in the Planner first");
-                    return;
-                }
-
-                // Everything past the preconditions -- schedule build, config injection,
-                // session create, event wiring, tracked RunAsync -- lives in
-                // SessionBootstrapper so this lambda routes only.
-                await SessionBootstrapper.BuildAndStartAsync(
-                    sp.GetRequiredService<ISessionFactory>(),
-                    appState, plannerState, sessionState, liveSessionState, profile,
-                    tracker, external, _timeProvider, logger, cts.Token);
-            });
-
-            bus.Subscribe<ConfirmAbortSessionSignal>(_ =>
-            {
-                liveSessionState.SessionCts?.Cancel();
-                liveSessionState.ShowAbortConfirm = false;
-                liveSessionState.NeedsRedraw = true;
-                appState.NeedsRedraw = true;
-            });
-
-            // ---------------------------------------------------------------
-            // Polar alignment signals
-            // ---------------------------------------------------------------
-
-            bus.Subscribe<StartPolarAlignmentSignal>(sig =>
-            {
-                if (!EnsureSessionIdle("Session is running \u2014 polar alignment unavailable")) return;
-                if (liveSessionState.PolarAlignmentCts is not null)
-                {
-                    Notify(NotificationSeverity.Warning, "Polar alignment already running");
-                    return;
-                }
-                if (appState.ActiveProfile?.Data is not { } profileData || profileData.OTAs.Length == 0)
-                {
-                    Notify(NotificationSeverity.Warning, "No profile / OTA configured");
-                    return;
-                }
-                if (appState.DeviceHub is not { } hub)
-                {
-                    Notify(NotificationSeverity.Warning, "Device hub not available");
-                    return;
-                }
-                if (!TryGetConnected<IMountDriver>(hub, profileData.Mount, "Mount", out var mount, "Mount not connected \u2014 connect a mount first")) return;
-                if (profileData.SiteLatitude is not { } lat || profileData.SiteLongitude is not { } lon)
-                {
-                    Notify(NotificationSeverity.Warning, "Site location not configured for this profile");
-                    return;
-                }
-                var solverFactory = sp.GetRequiredService<IPlateSolverFactory>();
-
-                // Build the capture source (guider or main-camera path) + site; the
-                // device-resolution + capture-source wiring lives in PolarAlignmentActions
-                // so this lambda routes only.
-                var built = PolarAlignmentActions.BuildCaptureSource(
-                    sig, profileData, hub, mount, liveSessionState,
-                    external, sp.GetRequiredService<ICelestialObjectDB>(), _timeProvider, logger);
-                if (built.Error is { } buildError)
-                {
-                    Notify(NotificationSeverity.Warning, buildError);
-                    return;
-                }
-                var source = built.Source!;
-                var activeGuider = built.ActiveGuider;
-
-                var site = PolarAlignmentActions.BuildSite(profileData, hub, lat, lon);
-
-                // Setup-panel path supplies the full configuration. Toolbar /
-                // legacy callers leave Configuration null and pin only
-                // DeltaRaDeg; fall back to Default for everything else.
-                var config = sig.Configuration ?? (PolarAlignmentConfiguration.Default with { RotationDeg = sig.DeltaRaDeg });
-
-                var polarCts = CancellationTokenSource.CreateLinkedTokenSource(cts.Token);
-                liveSessionState.PolarAlignmentCts = polarCts;
-                liveSessionState.Mode = LiveSessionMode.PolarAlign;
-                liveSessionState.PolarPhase = PolarAlignmentPhase.Idle;
-                liveSessionState.PolarStatusMessage = "Starting polar alignment\u2026";
-                liveSessionState.PolarPhaseAResult = null;
-                liveSessionState.LastPolarSolve = null;
-                liveSessionState.NeedsRedraw = true;
-                appState.NeedsRedraw = true;
-
-                tracker.Run(async () =>
-                {
-                    var session = new PolarAlignmentSession(
-                        external, mount, source, solverFactory,
-                        _timeProvider, logger, site, config);
-                    try
-                    {
-                        // If the guider was looping/calibrating/guiding from a prior session,
-                        // stop it cleanly first. PHD2's LoopAsync (used inside GuiderCaptureSource)
-                        // will refuse if the app is mid-calibration. Best-effort — failure here
-                        // is logged but doesn't fail the routine; the orchestrator's first frame
-                        // will surface the real problem with a useful message.
-                        if (activeGuider is not null)
-                        {
-                            try
-                            {
-                                await activeGuider.StopCaptureAsync(TimeSpan.FromSeconds(10), polarCts.Token);
-                            }
-                            catch (Exception ex)
-                            {
-                                logger.LogWarning(ex, "PolarAlignment: guider StopCaptureAsync before run failed");
-                            }
-                        }
-
-                        await PolarAlignmentActions.RunAsync(session, liveSessionState, logger, polarCts.Token);
-                    }
-                    catch (OperationCanceledException ex)
-                    {
-                        logger.LogInformation(ex, "Polar alignment routine cancelled");
-                    }
-                    catch (Exception ex)
-                    {
-                        logger.LogError(ex, "Polar alignment routine failed");
-                        liveSessionState.PolarPhase = PolarAlignmentPhase.Failed;
-                        liveSessionState.PolarStatusMessage = $"Polar alignment error: {ex.Message}";
-                        Notify(NotificationSeverity.Error, $"Polar alignment failed: {ex.Message}");
-                    }
-                    finally
-                    {
-                        // ReverseAxisBack / Park / LeaveInPlace runs here.
-                        liveSessionState.PolarPhase = PolarAlignmentPhase.RestoringMount;
-                        liveSessionState.NeedsRedraw = true;
-                        try { await session.DisposeAsync(); }
-                        catch (Exception ex) { logger.LogWarning(ex, "PolarAlignmentSession dispose failed"); }
-                        liveSessionState.PolarAlignmentCts = null;
-                        polarCts.Dispose();
-                        // Drop back into preview mode unless the user already swapped tabs.
-                        if (liveSessionState.Mode == LiveSessionMode.PolarAlign)
-                        {
-                            liveSessionState.Mode = LiveSessionMode.Preview;
-                        }
-                        liveSessionState.PolarPhase = PolarAlignmentPhase.Idle;
-                        liveSessionState.NeedsRedraw = true;
-                        appState.NeedsRedraw = true;
-                    }
-                }, "PolarAlignment");
-            });
-
-            bus.Subscribe<CancelPolarAlignmentSignal>(_ =>
-            {
-                // The Cancel button itself flips to an amber "Cancelling..." state
-                // while PolarAlignmentCts.IsCancellationRequested is true, and the
-                // phase pill carries the technical "RESTORING" badge as the mount
-                // reverses, so a third copy on the status line would be redundant.
-                liveSessionState.PolarAlignmentCts?.Cancel();
-                liveSessionState.NeedsRedraw = true;
-                appState.NeedsRedraw = true;
-            });
-
-            bus.Subscribe<NudgeFakeMountMisalignmentSignal>(sig =>
-            {
-                // Test path: nudges the simulated misalignment on a fake
-                // Skywatcher mount and shows the new value in the status line.
-                // For any real (or non-Skywatcher fake) mount, the keys can't
-                // turn physical knobs, so we surface a one-line hint instead
-                // of silently swallowing the input -- otherwise pressing arrows
-                // on a real-mount setup looks broken.
-                var profile = appState.ActiveProfile?.Data;
-                if (profile?.Mount is not { } mountUri || appState.DeviceHub is not { } hub)
-                {
-                    return;
-                }
-                if (!hub.TryGetConnectedDriver<IMountDriver>(mountUri, out var mount) || mount is null)
-                {
-                    return;
-                }
-                if (mount is TianWen.Lib.Devices.Fake.FakeSkywatcherMountDriver fake)
-                {
-                    fake.NudgeMisalignment(sig.DeltaAzArcmin, sig.DeltaAltArcmin);
-                    var (az, alt) = fake.CurrentMisalignment;
-                    liveSessionState.PolarStatusMessage =
-                        $"Sim misalignment: Az {az:+0.0;-0.0;0}', Alt {alt:+0.0;-0.0;0}'";
-                }
-                else
-                {
-                    liveSessionState.PolarStatusMessage = "Adjust the mount's alt/az knobs to refine";
-                }
-                liveSessionState.NeedsRedraw = true;
-                appState.NeedsRedraw = true;
-            });
-
-            bus.Subscribe<DonePolarAlignmentSignal>(_ =>
-            {
-                // Done is the same exit path as Cancel: stop the refine loop, let the
-                // session's DisposeAsync apply the configured OnDone behaviour
-                // (ReverseAxisBack by default).
-                liveSessionState.PolarAlignmentCts?.Cancel();
-                liveSessionState.PolarStatusMessage = "Restoring mount\u2026";
-                liveSessionState.NeedsRedraw = true;
-                appState.NeedsRedraw = true;
-            });
-
-            // ---------------------------------------------------------------
-            // Flats signals (on-demand flat capture -- LiveSessionMode.Flats)
-            // ---------------------------------------------------------------
-
-            bus.Subscribe<StartFlatsSignal>(async sig =>
-            {
-                if (!EnsureSessionIdle("Session is running \u2014 flats unavailable")) return;
-                if (liveSessionState.FlatsCts is not null)
-                {
-                    Notify(NotificationSeverity.Warning, "Flat run already running");
-                    return;
-                }
-                if (appState.ActiveProfile is not { Data: { } profileData } profile || profileData.OTAs.Length == 0)
-                {
-                    Notify(NotificationSeverity.Warning, "No profile / OTA configured");
-                    return;
-                }
-
-                // Site drives the mount sync + denorm stamp + sky-flat solar-altitude gate; NaN falls back
-                // to the mount's own site inside ConnectForFlatsAsync (matches the CLI path).
-                var siteLat = profileData.SiteLatitude ?? double.NaN;
-                var siteLon = profileData.SiteLongitude ?? double.NaN;
-
-                // Everything past the preconditions -- factory init, config injection, session create,
-                // event wiring, tracked RunFlatsOnlyAsync -- lives in FlatsBootstrapper so this lambda
-                // routes only (see CLAUDE.md "Signal Handler Pattern").
-                await FlatsBootstrapper.BuildAndStartAsync(
-                    sp.GetRequiredService<ISessionFactory>(),
-                    appState, sessionState, liveSessionState, profile,
-                    sig.Source, sig.FlatsPerFilter, siteLat, siteLon,
-                    tracker, _timeProvider, logger, cts.Token);
-            });
-
-            bus.Subscribe<CancelFlatsSignal>(_ =>
-            {
-                // The finaliser (close covers, warm, disconnect) still runs on cancel via RunFlatsOnlyAsync's
-                // finally block; the panel's Cancel button shows the amber "Cancelling..." state meanwhile.
-                liveSessionState.FlatsCts?.Cancel();
-                liveSessionState.NeedsRedraw = true;
-                appState.NeedsRedraw = true;
-            });
-
-            bus.Subscribe<RespondSessionPromptSignal>(sig =>
-            {
-                // Forward the user's Continue/Cancel to the session's awaiting prompt and drop the overlay.
-                if (liveSessionState.PendingPrompt is { } prompt)
-                {
-                    prompt.Respond(sig.Proceed);
-                    liveSessionState.PendingPrompt = null;
-                    liveSessionState.NeedsRedraw = true;
-                    appState.NeedsRedraw = true;
-                }
-            });
-
-            // ---------------------------------------------------------------
-            // Preview mode signals (camera preview, snapshot save, plate solve)
-            // ---------------------------------------------------------------
-
-            bus.Subscribe<TakePreviewSignal>(sig =>
-            {
-                if (!EnsureSessionIdle("Session is running \u2014 preview unavailable")) return;
-                if (appState.ActiveProfile?.Data is not { } previewData || sig.OtaIndex >= previewData.OTAs.Length)
-                {
-                    Notify(NotificationSeverity.Warning, "Invalid OTA index");
-                    return;
-                }
-                if (appState.DeviceHub is not { } hub) return;
-
-                var ota = previewData.OTAs[sig.OtaIndex];
-                if (!TryGetConnected<ICameraDriver>(hub, ota.Camera, "Camera", out var camera)) return;
-
-                // Resolve the OTA's other devices for per-capture FITS denorm. Mount is
-                // optional (preview can fire without one) but unlocks Target stamping
-                // and FakeCameraDriver synthetic-catalog rendering when present.
-                var (previewFocuser, previewFilterWheel, previewMount) =
-                    EquipmentActions.ResolveOtaCaptureDevices(hub, previewData, sig.OtaIndex);
-
-                // Mark capturing
-                if (sig.OtaIndex < liveSessionState.PreviewCapturing.Length)
-                {
-                    liveSessionState.PreviewCapturing[sig.OtaIndex] = true;
-                    liveSessionState.PreviewCaptureStart[sig.OtaIndex] = _timeProvider.GetUtcNow();
-                    liveSessionState.PreviewExposureDuration[sig.OtaIndex] = TimeSpan.FromSeconds(sig.ExposureSeconds);
-                }
-                appState.NeedsRedraw = true;
-
-                RunTracked($"PreviewCapture OTA{sig.OtaIndex}", "Preview failed", async ct =>
-                {
-                    // Stamp denorm fields before exposing -- shared with Session.Imaging
-                    // and polar alignment so the live preview's FITS headers (and the
-                    // FakeCameraDriver's synthetic-catalog rendering) match the other
-                    // capture paths exactly.
-                    await CameraExposureActions.StampDenormAsync(
-                        camera,
-                        ota.Name,
-                        ota.FocalLength,
-                        ota.Aperture,
-                        previewFocuser,
-                        previewFilterWheel,
-                        previewMount,
-                        targetName: previewMount is not null ? "Preview" : null,
-                        catalogDb: sp.GetRequiredService<ICelestialObjectDB>(),
-                        logger: logger,
-                        ct: ct).ConfigureAwait(false);
-
-                    var image = await LiveSessionActions.CaptureCameraPreviewAsync(
-                        camera,
-                        TimeSpan.FromSeconds(sig.ExposureSeconds),
-                        sig.Gain is { } g ? (short)g : null,
-                        sig.Binning,
-                        _timeProvider,
-                        ct);
-
-                    if (image is not null
-                        && sig.OtaIndex < liveSessionState.LastCapturedImages.Length)
-                    {
-                        // Release the previous slot's image before replacing -- otherwise
-                        // its ChannelBuffer ref never drops and the camera can't recycle
-                        // (mirrors the polar-refine onFrameCaptured leak fix).
-                        liveSessionState.LastCapturedImages[sig.OtaIndex]?.Release();
-                        liveSessionState.LastCapturedImages[sig.OtaIndex] = image;
-                        Notify(NotificationSeverity.Info, $"Preview captured: OTA {sig.OtaIndex + 1}");
-                    }
-                }, onFinally: () =>
-                {
-                    if (sig.OtaIndex < liveSessionState.PreviewCapturing.Length)
-                    {
-                        liveSessionState.PreviewCapturing[sig.OtaIndex] = false;
-                    }
-                    appState.NeedsRedraw = true;
-                }, cancelMessage: "Preview cancelled");
-            });
-
-            bus.Subscribe<SaveSnapshotSignal>(sig =>
-            {
-                if (liveSessionState.IsRunning) return;
-                if (sig.OtaIndex >= liveSessionState.LastCapturedImages.Length) return;
-                if (liveSessionState.LastCapturedImages[sig.OtaIndex] is not { } image)
-                {
-                    Notify(NotificationSeverity.Warning, "No preview image to save");
-                    return;
-                }
-
-                RunTracked("SaveSnapshot", "Snapshot failed", async _ =>
-                {
-                    var fileName = await LiveSessionActions.SaveSnapshotAsync(
-                        image, sig.OtaIndex, external, _timeProvider);
-                    Notify(NotificationSeverity.Info, $"Snapshot saved: {fileName}");
-                }, onFinally: () => appState.NeedsRedraw = true);
-            });
-
-            bus.Subscribe<PlateSolvePreviewSignal>(sig =>
-            {
-                if (liveSessionState.IsRunning) return;
-                if (sig.OtaIndex >= liveSessionState.LastCapturedImages.Length) return;
-                if (liveSessionState.LastCapturedImages[sig.OtaIndex] is not { } image)
-                {
-                    Notify(NotificationSeverity.Warning, "No preview image to solve");
-                    return;
-                }
-                // Drop duplicate clicks: if a solve is already running for this OTA,
-                // ignore. The button is rendered as "Solving…" with no click handler,
-                // but a stray hit before the redraw could still fire the signal.
-                if (sig.OtaIndex < liveSessionState.PreviewPlateSolving.Length
-                    && liveSessionState.PreviewPlateSolving[sig.OtaIndex])
-                {
-                    return;
-                }
-
-                if (sig.OtaIndex < liveSessionState.PreviewPlateSolving.Length)
-                {
-                    liveSessionState.PreviewPlateSolving[sig.OtaIndex] = true;
-                }
-                liveSessionState.NeedsRedraw = true;
-                appState.NeedsRedraw = true;
-
-                RunTracked("PreviewPlateSolve", "Plate solve error", async ct =>
-                {
-                    appState.StatusMessage = "Plate solving\u2026";
-                    appState.NeedsRedraw = true;
-
-                    // Solve orchestration (search-origin derivation + result-to-message
-                    // mapping) lives in LiveSessionActions so this lambda routes only.
-                    var (result, message, solved) = await LiveSessionActions.SolvePreviewFrameAsync(
-                        sp.GetRequiredService<IPlateSolverFactory>(), image, ct);
-                    liveSessionState.PreviewPlateSolveResult = result;
-                    Notify(solved ? NotificationSeverity.Info : NotificationSeverity.Warning, message);
-                }, onFinally: () =>
-                {
-                    if (sig.OtaIndex < liveSessionState.PreviewPlateSolving.Length)
-                    {
-                        liveSessionState.PreviewPlateSolving[sig.OtaIndex] = false;
-                    }
-                    liveSessionState.NeedsRedraw = true;
-                    appState.NeedsRedraw = true;
-                });
-            });
-
-            bus.Subscribe<JogFocuserSignal>(sig =>
-            {
-                if (!TryResolveIdleOtaFocuser(sig.OtaIndex, out var focuser)) return;
-
-                RunTracked($"JogFocuser OTA{sig.OtaIndex}", "Focuser jog failed", async ct =>
-                {
-                    var targetPos = await LiveSessionActions.JogFocuserAsync(focuser, sig.Steps, ct);
-                    Notify(NotificationSeverity.Info, $"Focuser \u2192 {targetPos}");
-                }, onFinally: () => appState.NeedsRedraw = true);
-            });
-
-            // Live planetary capture: route Start/Stop to the shared PlanetaryCaptureController (the capture
-            // loop + rolling-window stack live there; this only resolves the camera + configures the ROI).
-            var planetaryCapture = sp.GetRequiredService<PlanetaryCaptureController>();
-
-            bus.Subscribe<StartVideoCaptureSignal>(sig =>
-            {
-                if (liveSessionState.IsRunning) return;
-                if (appState.ActiveProfile?.Data is not { OTAs: var otas } || sig.OtaIndex >= otas.Length) return;
-                if (appState.DeviceHub is not { } hub) return;
-
-                var ota = otas[sig.OtaIndex];
-                if (!TryGetConnected<ICameraDriver>(hub, ota.Camera, "Camera", out var camera, "Connect a camera to start a planetary capture")) return;
-
-                var (roiW, roiH) = PlanetaryCaptureActions.ConfigureRoi(camera, sig.RoiWidth, sig.RoiHeight);
-
-                // Wire the coupled mount + OTA pixel scale for the COM recenter loop's coarse mount-jog
-                // fallback (Phase C). No-op when no mount is connected or the scale is unknown (NaN) -- the
-                // recenter loop then stays ROI-only.
-                IMountDriver? recenterMount = null;
-                if (appState.ActiveProfile?.Data is { } capturePdata
-                    && capturePdata.Mount is { Scheme: not "none" } mountUri
-                    && hub.TryGetConnectedDriver<IMountDriver>(mountUri, out var rcMount) && rcMount is not null)
-                {
-                    recenterMount = rcMount;
-                }
-                planetaryCapture.AttachMount(
-                    recenterMount, CoordinateUtils.PixelScaleArcsec(camera.PixelSizeX, ota.FocalLength));
-
-                // Bind the capture's lifetime to the app shutdown token: quitting cancels it (its loops poll
-                // the token), so the camera is released without an imperative Stop() in the quit path.
-                planetaryCapture.Start(camera,
-                    new VideoCaptureOptions(TimeSpan.FromMilliseconds(sig.ExposureMs), sig.Gain), shutdownToken);
-                // Planetary capture is now a Live Session mode (not a standalone tab): show it there.
-                liveSessionState.Mode = LiveSessionMode.Planetary;
-                appState.ActiveTab = GuiTab.LiveSession;
-                Notify(NotificationSeverity.Info, $"Planetary capture started ({roiW}x{roiH}, {sig.ExposureMs:F0} ms)");
-            });
-
-            bus.Subscribe<StopVideoCaptureSignal>(_ =>
-            {
-                planetaryCapture.Stop();
-                Notify(NotificationSeverity.Info, "Planetary capture stopped");
-            });
-
-            // Manual mount nudge (planetary panel coarse-recenter buttons) -> the same pulse-guide actuator the
-            // COM recenter loop uses. Mirrors the focuser-jog route: resolve the connected mount, pulse on the
-            // tracker. Gated on no running session + a pulse-guide-capable mount.
-            bus.Subscribe<JogMountSignal>(sig =>
-            {
-                if (liveSessionState.IsRunning) return;
-                if (appState.ActiveProfile?.Data is not { } pdata) return;
-                if (pdata.Mount is not { Scheme: not "none" } mountUri) return;
-                if (appState.DeviceHub is not { } hub) return;
-                if (!hub.TryGetConnectedDriver<IMountDriver>(mountUri, out var mount) || mount is null) return;
-
-                RunTracked($"JogMount {sig.Direction}", "Mount jog failed", async ct =>
-                {
-                    await MountActions.PulseGuideArcsecAsync(
-                        mount, sig.Direction, sig.Arcsec, _timeProvider, logger: logger, cancellationToken: ct);
-                    Notify(NotificationSeverity.Info, $"Mount nudge {sig.Direction} {sig.Arcsec:F0} arcsec");
-                }, onFinally: () => appState.NeedsRedraw = true);
-            });
-
-            bus.Subscribe<GotoFocuserSignal>(sig =>
-            {
-                if (!TryResolveIdleOtaFocuser(sig.OtaIndex, out var focuser)) return;
-
-                RunTracked($"GotoFocuser OTA{sig.OtaIndex}", "Focuser goto failed", async ct =>
-                {
-                    await focuser.BeginMoveAsync(sig.TargetPosition, ct);
-                    Notify(NotificationSeverity.Info, $"Focuser \u2192 {sig.TargetPosition}");
-                }, onFinally: () => appState.NeedsRedraw = true);
-            });
-
-            // ---------------------------------------------------------------
-            // Local helpers captured by closures above
-            // ---------------------------------------------------------------
-            void CommitSuggestion(string suggestion)
-            {
-                plannerState.SearchInput.Text = suggestion;
-                plannerState.SearchInput.CursorPos = suggestion.Length;
-                plannerState.Suggestions.Clear();
-                plannerState.SuggestionIndex = -1;
-                plannerState.LastSuggestionQuery = suggestion;
-
-                if (appState.ActiveProfile is not null)
-                {
-                    var transform = TransformFactory.FromProfile(appState.ActiveProfile, _timeProvider, out _);
-                    if (transform is not null)
-                    {
-                        var db = sp.GetRequiredService<TianWen.Lib.Astrometry.Catalogs.ICelestialObjectDB>();
-                        var resultIdx = PlannerActions.CommitSuggestion(plannerState, db, transform, suggestion, plannerState.Comets);
-                        if (resultIdx >= 0)
-                        {
-                            plannerState.SelectedTargetIndex = resultIdx;
-                            OnPlannerEnsureVisible?.Invoke(resultIdx);
-                        }
-                    }
-                }
-                appState.NeedsRedraw = true;
-            }
+            // Subscription groups live in the by-area partials (AppSignalHandler.Planner.cs /
+            // .SkyMap.cs / .Equipment.cs / .LiveSession.cs / .Polar.cs / .Flats.cs). Order is
+            // load-bearing: SignalBus invokes subscribers in registration order.
+            SubscribePlannerSearch(bus);
+            SubscribeSkyMap(bus);
+            SubscribeEquipmentTextInputs(bus);
+            SubscribeEquipmentActions(bus);
+            SubscribeScheduleBuilding(bus);
+            SubscribeLiveSession(bus);
+            SubscribePolarAlignment(bus);
+            SubscribeFlats(bus);
+            SubscribePreview(bus, shutdownToken);
 
             // Store autocomplete cache setter as a public action
-            SetAutoCompleteCache = cache => autoCompleteCache = cache;
+            SetAutoCompleteCache = cache => _autoCompleteCache = cache;
         }
 
         // -----------------------------------------------------------------------------
